stm32/rcc: unify f0, f1, f3.
