+++ embassy-stm32/src/rcc/f247.rs
